move /shared
