+++ library/src/app/components/price-list/price-list.component.spec.ts
@@ -10,27 +10,27 @@ import { PriceListComponent } from './price-list.component';
 import { PricesService } from '@cybrid/cybrid-api-bank-angular';
 
 // Services
-import { AuthService } from '../../../../../src/shared/services/auth/auth.service';
-import { EventService } from '../../../../../src/shared/services/event/event.service';
-import { ErrorService } from '../../../../../src/shared/services/error/error.service';
-import { ConfigService } from '../../../../../src/shared/services/config/config.service';
+import { AuthService } from '../../../shared/services/auth/auth.service';
+import { EventService } from '../../../shared/services/event/event.service';
+import { ErrorService } from '../../../shared/services/error/error.service';
+import { ConfigService } from '../../../shared/services/config/config.service';
 
 // Modules
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { SharedModule } from '../../../../../src/shared/modules/shared.module';
+import { SharedModule } from '../../../shared/modules/shared.module';
 import { ReactiveFormsModule } from '@angular/forms';
 
 // Testing
 import { of, throwError } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import { AssetService } from '../../../../../src/shared/services/asset/asset.service';
+import { AssetService } from '../../../shared/services/asset/asset.service';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import { HttpLoaderFactory } from '../../modules/library.module';
-import { TestConstants } from '../../../../../src/shared/constants/test.constants';
+import { TestConstants } from '../../../shared/constants/test.constants';
 import { RouterTestingModule } from '@angular/router/testing';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
-import { RoutingService } from '../../../../../src/shared/services/routing/routing.service';
+import { RoutingService } from '../../../shared/services/routing/routing.service';
 
 describe('ListComponent', () => {
   let MockAuthService = jasmine.createSpyObj('AuthService', [
