+++ library/src/app/components/trade-summary/trade-summary.component.ts
@@ -2,17 +2,17 @@ import { Component, Inject, OnInit } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { TradeBankModel, TradesService } from '@cybrid/cybrid-api-bank-angular';
 import { BehaviorSubject, catchError, map, of, Subject } from 'rxjs';
-import { ConfigService } from '../../../../../src/shared/services/config/config.service';
+import { ConfigService } from '../../../shared/services/config/config.service';
 import {
   CODE,
   EventService,
   LEVEL
-} from '../../../../../src/shared/services/event/event.service';
-import { ErrorService } from '../../../../../src/shared/services/error/error.service';
+} from '../../../shared/services/event/event.service';
+import { ErrorService } from '../../../shared/services/error/error.service';
 import { TranslatePipe } from '@ngx-translate/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { Asset } from '../../../../../src/shared/services/asset/asset.service';
-import { RoutingService } from '../../../../../src/shared/services/routing/routing.service';
+import { Asset } from '../../../shared/services/asset/asset.service';
+import { RoutingService } from '../../../shared/services/routing/routing.service';
 
 interface DialogData {
   model: TradeBankModel;
@@ -25,7 +25,7 @@ interface DialogData {
   templateUrl: './trade-summary.component.html',
   styleUrls: [
     './trade-summary.component.scss',
-    '../../../../../src/shared/styles/global.scss'
+    '../../../shared/styles/global.scss'
   ]
 })
 export class TradeSummaryComponent implements OnInit {
