+++ library/src/app/components/trade/trade.component.spec.ts
@@ -12,11 +12,11 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import { HttpLoaderFactory } from '../../modules/library.module';
 import { HttpClient } from '@angular/common/http';
-import { AssetPipe } from '../../../../../src/shared/pipes/asset/asset.pipe';
+import { AssetPipe } from '../../../shared/pipes/asset/asset.pipe';
 import { ActivatedRoute } from '@angular/router';
 import { of, throwError } from 'rxjs';
-import { TestConstants } from '../../../../../src/shared/constants/test.constants';
-import { AssetService } from '../../../../../src/shared/services/asset/asset.service';
+import { TestConstants } from '../../../shared/constants/test.constants';
+import { AssetService } from '../../../shared/services/asset/asset.service';
 import { PricesService } from '@cybrid/cybrid-api-bank-angular';
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { MatInputModule } from '@angular/material/input';
@@ -24,13 +24,13 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatSelectModule } from '@angular/material/select';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { EventService } from '../../../../../src/shared/services/event/event.service';
-import { ErrorService } from '../../../../../src/shared/services/error/error.service';
-import { ConfigService } from '../../../../../src/shared/services/config/config.service';
+import { EventService } from '../../../shared/services/event/event.service';
+import { ErrorService } from '../../../shared/services/error/error.service';
+import { ConfigService } from '../../../shared/services/config/config.service';
 import { MatDialog } from '@angular/material/dialog';
-import { QuoteService } from '../../../../../src/shared/services/quote/quote.service';
+import { QuoteService } from '../../../shared/services/quote/quote.service';
 import { TradeConfirmComponent } from '../trade-confirm/trade-confirm.component';
-import { RoutingService } from '../../../../../src/shared/services/routing/routing.service';
+import { RoutingService } from '../../../shared/services/routing/routing.service';
 
 describe('TradeComponent', () => {
   let component: TradeComponent;
