+++ library/src/shared/i18n/fr.ts
