+++ library/src/shared/pipes/asset/asset.pipe.ts
