+++ library/src/shared/pipes/asset/mock-asset.pipe.ts
