+++ library/src/shared/pipes/truncate/tuncate.pipe.ts
