+++ library/src/shared/services/config/config.service.ts
