+++ library/src/shared/services/error/error.service.ts
