+++ library/src/shared/services/quote/quote.service.spec.ts
