+++ library/src/shared/services/quote/quote.service.ts
