+++ library/src/shared/styles/theme.scss
@@ -1,4 +1,4 @@
-@use '@angular/material' as mat;
+@use '../../../../node_modules/@angular/material/index' as mat;
 @include mat.core();
 
 $dark-primary-text: rgba(black, 0.87);
