+++ library/src/shared/utility/compare-object.ts
