+++ library/src/shared/utility/strip-format.ts
