+++ library/src/shared/utility/symbol-build.ts
