+++ library/src/shared/utility/symbol-split.ts
