+++ package.json
@@ -5,11 +5,11 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build --output-hashing=none",
-    "package:styles": "echo 'Bundling styles into Javascript' && npx webpack --mode=production --config webpack.styles.config.js",
-    "package:polyfills": "echo 'Copying polyfills' && cp dist/library/polyfills.js dist/library/cybrid-sdk-ui.polyfills.js",
-    "package:main": "echo 'Bundling main.js and runtime.js' && cat dist/library/runtime.js dist/library/main.js dist/library/cybrid-sdk-ui.styles.js > dist/library/cybrid-sdk-ui.min.js",
-    "package:extras": "echo 'Copying README.md and package.json' && cp library/README.md dist/library && cp library/package.json dist/library",
-    "package:library": "echo 'Building and packaging library' && ng build library && npm run package:styles && npm run package:polyfills && npm run package:main && npm run package:extras && npx webpack --mode=production --config webpack.main.config.js",
+    "package:styles": "echo 'Bundling styles into Javascript \n' && npx webpack --mode=production --config webpack.styles.config.js",
+    "package:polyfills": "echo 'Copying polyfills \n' && cp dist/library/polyfills.js dist/library/cybrid-sdk-ui.polyfills.js",
+    "package:main": "echo 'Bundling main.js and runtime.js \n' && cat dist/library/runtime.js dist/library/main.js dist/library/cybrid-sdk-ui.styles.js > dist/library/cybrid-sdk-ui.min.js",
+    "package:extras": "echo 'Copying README.md and package.json \n' && cp library/README.md dist/library && cp library/package.json dist/library",
+    "package:library": "echo 'Building and packaging library \n' && ng build library && npm run package:styles && npm run package:polyfills && npm run package:main && npm run package:extras && npx webpack --mode=production --config webpack.main.config.js",
     "publish:library": "cd dist/library && npm publish --access public",
     "watch": "ng build --watch --configuration development",
     "test": "echo 'clearing angular cache...' && rm -rf .angular && ng run demo:test && echo 'clearing angular cache...' && rm -rf .angular && ng run library:test",
