+++ src/styles.scss
@@ -1,6 +1,6 @@
 @use '@angular/material' as mat;
 @import url('https://fonts.googleapis.com/icon?family=Material+Icons');
-@import '/src/shared/styles/theme.scss';
+@import '/library/src/shared/styles/theme.scss';
 
 html,
 body {
